Integrate SymbolInformationOps into Scala3#SymbolOps
